extract compute.ts
